Create autounattend.xml
For the create a Windows Templeate on OSP Virtualization: With the options filled out, we want to click on the Customize VirtualMachine button at the bottom to continue configuring our Template. On the Customize and create VirtualMachine screen, click on the Scripts tab, and then scroll down to the Sysprep section and click on the Edit button. A new window will pop up for you to create Sysprep actions for your new template.