Create Cloud-init.yaml
On the OSP Virtualization set: Next click on the Scripts tab at the top, and in the section called Cloud-init click the Edit button. When the Cloud-init dialog opens, click the radio button to Configure via: Script then replace the YAML with the following YAML snippet. Click the Save button, followed by the Apply button. Now click on the Virtualization → Catalog menu on the left, select the Template catalog option and click on User templates. You should see our created template available as a tile.